+++ packages/synchronizer/src/handlers/apiHandler.ts
@@ -2,6 +2,7 @@ import normalizeUrl from 'normalize-url';
 import fetch from 'node-fetch';
 import {SuppressionStatus} from '@monokle/types';
 import {DEFAULT_API_URL} from '../constants.js';
+import type {TokenInfo} from './storageHandlerAuth.js';
 
 const getUserQuery = `
   query getUser {
@@ -161,20 +162,20 @@ export class ApiHandler {
     return normalizeUrl(this._apiUrl);
   }
 
-  async getUser(accessToken: string): Promise<ApiUserData | undefined> {
-    return this.queryApi(getUserQuery, accessToken);
+  async getUser(tokenInfo: TokenInfo): Promise<ApiUserData | undefined> {
+    return this.queryApi(getUserQuery, tokenInfo);
   }
 
-  async getProject(slug: string, accessToken: string): Promise<ApiProjectData | undefined> {
-    return this.queryApi(getProjectQuery, accessToken, {slug});
+  async getProject(slug: string, tokenInfo: TokenInfo): Promise<ApiProjectData | undefined> {
+    return this.queryApi(getProjectQuery, tokenInfo, {slug});
   }
 
-  async getPolicy(slug: string, accessToken: string): Promise<ApiPolicyData | undefined> {
-    return this.queryApi(getPolicyQuery, accessToken, {slug});
+  async getPolicy(slug: string, tokenInfo: TokenInfo): Promise<ApiPolicyData | undefined> {
+    return this.queryApi(getPolicyQuery, tokenInfo, {slug});
   }
 
-  async getSuppressions(repositoryId: string, accessToken: string): Promise<ApiSuppressionsData | undefined> {
-    return this.queryApi(getSuppressionsQuery, accessToken, {repositoryId});
+  async getSuppressions(repositoryId: string, tokenInfo: TokenInfo): Promise<ApiSuppressionsData | undefined> {
+    return this.queryApi(getSuppressionsQuery, tokenInfo, {repositoryId});
   }
 
   generateDeepLink(path: string) {
@@ -189,20 +190,9 @@ export class ApiHandler {
     return normalizeUrl(`${this.apiUrl}/${path}`);
   }
 
-  private async queryApi<OUT>(query: string, token: string, variables = {}): Promise<OUT | undefined> {
+  private async queryApi<OUT>(query: string, tokenInfo: TokenInfo, variables = {}): Promise<OUT | undefined> {
     const apiEndpointUrl = normalizeUrl(`${this.apiUrl}/graphql`);
-
-    const response = await fetch(apiEndpointUrl, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
-      },
-      body: JSON.stringify({
-        query,
-        variables,
-      }),
-    });
+    const response = await this.sendRequest(apiEndpointUrl, tokenInfo, query, variables);
 
     if (!response.ok) {
       throw new Error(
@@ -210,6 +200,44 @@ export class ApiHandler {
       );
     }
 
-    return response.json() as Promise<OUT>;
+    const responseJson = await response.json();
+
+    if (responseJson?.errors?.length > 0) {
+      const error = responseJson.errors[0];
+      const msg = error.message;
+      const code = error.extensions?.code;
+
+      if (msg === 'Unauthorized' || code === 'UNAUTHENTICATED') {
+        throw new Error(
+          `Unauthorized error. Make sure that valid auth credentials were used. Cannot fetch data from ${apiEndpointUrl}.`
+        );
+      }
+
+      throw new Error(`${msg} error (code: ${code}). Cannot fetch data from ${apiEndpointUrl}.`);
+    }
+
+    return responseJson as Promise<OUT>;
+  }
+
+  private async sendRequest(apiEndpointUrl: string, tokenInfo: TokenInfo, query: string, variables = {}) {
+    return fetch(apiEndpointUrl, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: this.formatAuthorizationHeader(tokenInfo),
+      },
+      body: JSON.stringify({
+        query,
+        variables,
+      }),
+    });
+  }
+
+  private formatAuthorizationHeader(tokenInfo: TokenInfo) {
+    const tokenType =
+      tokenInfo?.tokenType?.toLowerCase() === 'bearer' || tokenInfo?.tokenType?.toLowerCase() === 'apikey'
+        ? tokenInfo.tokenType
+        : 'Bearer';
+    return `${tokenType} ${tokenInfo.accessToken}`;
   }
 }