+++ packages/validation/src/validators/pod-security-standards/rules/PSS106-host-ports.ts
@@ -45,4 +45,5 @@ export const hostPorts = defineRule({
       });
     });
   },
+  fix: undefined, // Autofix makes no sense, they need to decide on another kind of volume.
 });