+++ packages/synchronizer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @monokle/synchronizer
 
+## 0.9.1
+
+### Patch Changes
+
+- 7838f08: Added more explicit GraphQL API error handling
+
+## 0.9.0
+
+### Minor Changes
+
+- 8d03837: Require token object instead of string for API queries
+
 ## 0.8.0
 
 ### Minor Changes