+++ packages/validation/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @monokle/validation
 
+## 0.30.1
+
+### Patch Changes
+
+- 739ab86: Polish autofix
+
+## 0.30.0
+
+### Minor Changes
+
+- c317d85: add pss autofixes
+
 ## 0.29.1
 
 ### Patch Changes